+++ Source/MQTTnet.Extensions.WebSocket4Net/MqttFactoryExtensions.cs
@@ -10,7 +10,10 @@ namespace MQTTnet.Extensions.WebSocket4Net
     {
         public static MqttFactory UseWebSocket4Net(this MqttFactory mqttFactory)
         {
-            if (mqttFactory == null) throw new ArgumentNullException(nameof(mqttFactory));
+            if (mqttFactory == null)
+            {
+                throw new ArgumentNullException(nameof(mqttFactory));
+            }
 
             return mqttFactory.UseClientAdapterFactory(new WebSocket4NetMqttClientAdapterFactory());
         }